Pytorch_Retinaface-master/requirements.txt
torch==1.13.1+cu116
opencv-python==4.7.0.72
numpy==1.21.6
torchvision==0.14.1+cu116
pyqt5==5.15.9
pillow==9.1.1
cython==0.29.34
tqdm==4.65.0
scipy==1.7.3
ipython==7.34.0